debug compilation

+++ src/ExpDiv.cpp
@@ -5,7 +5,7 @@ using namespace std;
 	ExpDiv::ExpDiv(Exp * g, Exp * d) : ExpBin(g,d) { } 
 	ExpDiv::~ExpDiv() { }
 
-	int ExpDiv::operation(int g, int d) {
+	double ExpDiv::operation(double g, double d) {
 		if(d==0) std::cerr << "Division par zero impossible" << endl;
 	   return g/d;
 	}
